auto-pdftotext when reading pdf files
When *.pdf is read, automatically open a non-file buffer with same name
but with .txt extension, then filter the pdf through pdftotext and read
it into the non-file buffer. This is just a small nicety if you ever
open a pdf or need to view it without a GUI pdf viewer.